+++ FightObjects/AllFights.cs
@@ -8,8 +8,8 @@ namespace BecomeSifu.FightObjects
     public class AllFights
     {
         public int Wins { get; set; }
-        public ulong Health { get; set; }
-        public ulong Attack { get; set; }
+        public decimal Health { get; set; }
+        public decimal Attack { get; set; }
         public bool IsActive { get; set; }
         public string FightName { get; set; }
         public SolidColorBrush Background { get; set; }
