+++ FightObjects/BTournament.cs
@@ -10,16 +10,16 @@ namespace BecomeSifu.FightObjects
         public BTournament()
         {
             Wins = 0;
-            Health = ((ulong)Wins + 1) * 100;
-            Attack = ((ulong)Wins + 1) * 1000;
+            Health = ((decimal)Wins + 1) * 100;
+            Attack = ((decimal)Wins + 1) * 1000;
             FightName = "Tournament Bout";
             Background = new SolidColorBrush(Colors.IndianRed);
         }
         public BTournament(int wins)
         {
             Wins = wins;
-            Health = ((ulong)Wins + 1) * 100;
-            Attack = ((ulong)Wins + 1) * 1000;
+            Health = ((decimal)Wins + 1) * 100;
+            Attack = ((decimal)Wins + 1) * 1000;
         }
 
         private BTournament Won(int win)
