+++ FightObjects/IFights.cs
@@ -9,8 +9,8 @@ namespace BecomeSifu.FightObjects
     public interface IFights
     {
         int Wins { get; set; }
-        ulong Health { get; set; }
-        ulong Attack { get; set; }
+        decimal Health { get; set; }
+        decimal Attack { get; set; }
         bool IsActive { get; set; }
         string FightName { get; }
         SolidColorBrush Background { get; }
